Allow replacing the shipment serializer entirely
