+++ lib/solidus_shipstation.rb
@@ -7,7 +7,7 @@ require 'solidus_shipstation/api/request_runner'
 require 'solidus_shipstation/api/client'
 require 'solidus_shipstation/api/request_error'
 require 'solidus_shipstation/api/rate_limited_error'
-require 'solidus_shipstation/api/serializer'
+require 'solidus_shipstation/api/shipment_serializer'
 require 'solidus_shipstation/api/threshold_verifier'
 require 'solidus_shipstation/configuration'
 require 'solidus_shipstation/version'
