+++ lib/solidus_shipstation/configuration.rb
@@ -12,7 +12,7 @@ module SolidusShipstation
       :api_batch_size,
       :api_sync_threshold,
       :api_store_id,
-      :custom_api_params,
+      :api_shipment_serializer,
       :api_username,
       :api_password,
       :error_handler,
@@ -21,10 +21,12 @@ module SolidusShipstation
     def initialize
       @api_batch_size = 100
       @api_sync_threshold = 7.days
-      @custom_api_params = ->(_shipment) { {} }
       @error_handler = ->(_error, _extra = {}) {
         Rails.logger.error "#{error.inspect} (#{extra.inspect})"
       }
+      @api_shipment_serializer = proc do |shipment|
+        SolidusShipstation::Api::ShipmentSerializer.new.call(shipment)
+      end
     end
   end
