+++ spec/lib/solidus_shipstation/api/client_spec.rb
@@ -3,14 +3,19 @@ RSpec.describe SolidusShipstation::Api::Client do
     it 'generates a client from the configuration' do
       request_runner = instance_double('SolidusShipstation::Api::RequestRunner')
       error_handler = instance_spy('Proc')
+      shipment_serializer = instance_spy('SolidusShipstation::Api::Serializer')
       allow(SolidusShipstation::Api::RequestRunner).to receive(:from_config).and_return(request_runner)
-      allow(SolidusShipstation.config).to receive(:error_handler).and_return(error_handler)
+      allow(SolidusShipstation.config).to receive_messages(
+        error_handler: error_handler,
+        api_shipment_serializer: shipment_serializer,
+      )
 
       client = described_class.from_config
 
       expect(client).to have_attributes(
         request_runner: request_runner,
         error_handler: error_handler,
+        shipment_serializer: shipment_serializer,
       )
     end
   end
@@ -18,11 +23,13 @@ RSpec.describe SolidusShipstation::Api::Client do
   describe '#bulk_create_orders' do
     it 'calls the bulk order creation endpoint' do
       request_runner = instance_spy('SolidusShipstation::Api::RequestRunner')
-      client = build_client(request_runner: request_runner)
       shipment = build_stubbed(:shipment)
       serialized_shipment = { 'key' => 'value' }
-      stub_serializer(shipment, serialized_shipment)
 
+      client = build_client(
+        request_runner: request_runner,
+        shipment_serializer: stub_shipment_serializer(shipment => serialized_shipment),
+      )
       client.bulk_create_orders([shipment])
 
       expect(request_runner).to have_received(:call).with(
@@ -34,14 +41,18 @@ RSpec.describe SolidusShipstation::Api::Client do
 
     it 'does not fail for serialization errors' do
       request_runner = instance_spy('SolidusShipstation::Api::RequestRunner')
-      client = build_client(request_runner: request_runner)
       successful_shipment = build_stubbed(:shipment)
+      serialized_shipment = { 'key' => 'value' }
       failing_shipment = build_stubbed(:shipment)
       error = RuntimeError.new('Failed to serialize shipment!')
-      serialized_shipment = { 'key' => 'value' }
-        stub_serializer(successful_shipment, serialized_shipment)
-        stub_serializer(failing_shipment, error)
 
+      client = build_client(
+        request_runner: request_runner,
+        shipment_serializer: stub_shipment_serializer(
+          successful_shipment => serialized_shipment,
+          failing_shipment => error,
+        )
+      )
       client.bulk_create_orders([failing_shipment, successful_shipment])
 
       expect(request_runner).to have_received(:call).with(
@@ -53,11 +64,13 @@ RSpec.describe SolidusShipstation::Api::Client do
 
     it 'reports any serialization errors to the error handler' do
       error_handler = instance_spy('Proc')
-      client = build_client(error_handler: error_handler)
       shipment = build_stubbed(:shipment)
       error = RuntimeError.new('Failed to serialize shipment!')
-      stub_serializer(shipment, error)
 
+      client = build_client(
+        shipment_serializer: stub_shipment_serializer(shipment => error),
+        error_handler: error_handler,
+      )
       client.bulk_create_orders([shipment])
 
       expect(error_handler).to have_received(:call).with(error, shipment: shipment)
@@ -65,10 +78,14 @@ RSpec.describe SolidusShipstation::Api::Client do
 
     it 'skips the API call if all shipments failed serialization' do
       request_runner = instance_spy('SolidusShipstation::Api::RequestRunner')
-      client = build_client(request_runner: request_runner)
       failing_shipment = build_stubbed(:shipment)
-      stub_serializer(failing_shipment, RuntimeError.new('Failed to serialize shipment!'))
 
+      client = build_client(
+        shipment_serializer: stub_shipment_serializer(
+          failing_shipment => RuntimeError.new('Failed to serialize shipment!'),
+        ),
+        request_runner: request_runner,
+      )
       client.bulk_create_orders([failing_shipment])
 
       expect(request_runner).not_to have_received(:call)
@@ -81,16 +98,23 @@ RSpec.describe SolidusShipstation::Api::Client do
     described_class.new({
       request_runner: instance_spy('SolidusShipstation::Api::RequestRunner'),
       error_handler: instance_spy('Proc'),
+      shipment_serializer: stub_shipment_serializer,
     }.merge(options))
   end
 
-  def stub_serializer(shipment, result_or_error)
-    stub = allow(SolidusShipstation::Api::Serializer).to(receive(:serialize_shipment).with(shipment))
+  def stub_shipment_serializer(results_map = {})
+    serializer = class_spy('SolidusShipstation::Api::Serializer')
 
-    if result_or_error.is_a?(Hash)
-      stub.and_return(result_or_error)
-    else
-      stub.and_raise(result_or_error)
+    results_map.each_pair do |shipment, result_or_error|
+      stub = allow(serializer).to receive(:call).with(shipment)
+
+      if result_or_error.is_a?(Hash)
+        stub.and_return(result_or_error)
+      else
+        stub.and_raise(result_or_error)
+      end
     end
+
+    serializer
   end
 end
